requirements.txt
numpy
Pillow
https://github.com/pyglet/pyglet/archive/master.zip
jupyterlab
matplotlib
nose
nose-timer
openpyxl
kaitaistruct
profilehooks
numba